drf-backend/requirements.txt
Django==4.1
djangorestframework==3.13.1
djoser==2.1.0
psycopg2-binary==2.9.3
django-cors-headers==3.12
django-environ==0.9.0
Pillow==9.2.0
redis==4.3.4
celery==5.2.7
django-celery-email==3.0.0
flower==1.2.0
tornado==6.2
watchdog==2.1.9